+++ cognite/replicator/time_series.py
@@ -91,6 +91,7 @@ def copy_ts(
     project_src: str,
     runtime: int,
     client: CogniteClient,
+    dst_ts: List[TimeSeries],
 ):
     """
     Creates/updates time series objects and then attempts to create and update these time series in the destination.
@@ -102,6 +103,7 @@ def copy_ts(
         project_src: The name of the project the object is being replicated from.
         runtime: The timestamp to be used in the new replicated metadata.
         client: The client corresponding to the destination project.
+        dst_ts: List of timeseries in the destination - Will be used for comparison if current timeseries where not copied by the replicator
     """
     logging.info(f"Starting to replicate {len(src_ts)} time series.")
     for ts in src_ts:  # for unlinkable time_series, remove asset id to avoid insertion error
@@ -109,7 +111,14 @@ def copy_ts(
             ts.asset_id = None
 
     create_ts, update_ts, unchanged_ts = replication.make_objects_batch(
-        src_ts, src_id_dst_ts, src_dst_ids_assets, create_time_series, update_time_series, project_src, runtime
+        src_ts,
+        src_id_dst_ts,
+        src_dst_ids_assets,
+        create_time_series,
+        update_time_series,
+        project_src,
+        runtime,
+        dst_ts=dst_ts,
     )
 
     logging.info(f"Creating {len(create_ts)} new time series and updating {len(update_ts)} existing time series.")
@@ -123,6 +132,9 @@ def copy_ts(
         logging.info(f"Updating {len(update_ts)} time series.")
         updated_ts = replication.retry(client.time_series.update, update_ts)
         logging.info(f"Successfully updated {len(updated_ts)} time series.")
+
+    if unchanged_ts:
+        logging.info(f"{len(unchanged_ts)} time series will not be changed.")
 
 
 def replicate(
@@ -158,9 +170,9 @@ def replicate(
     project_dst = client_dst.config.project
 
     if target_external_ids:
-        ts_src = client_src.time_series.retrieve_multiple(external_ids=target_external_ids)
+        ts_src = client_src.time_series.retrieve_multiple(external_ids=target_external_ids, ignore_unknown_ids=True)
         try:
-            ts_dst = client_dst.time_series.retrieve_multiple(external_ids=target_external_ids)
+            ts_dst = client_dst.time_series.retrieve_multiple(external_ids=target_external_ids, ignore_unknown_ids=True)
         except CogniteNotFoundError:
             ts_dst = TimeSeriesList([])
     else:
@@ -210,6 +222,7 @@ def replicate(
             project_src=project_src,
             replicated_runtime=replicated_runtime,
             client=client_dst,
+            dst_ts=ts_dst,
         )
     else:
         copy_ts(
@@ -219,6 +232,7 @@ def replicate(
             project_src=project_src,
             runtime=replicated_runtime,
             client=client_dst,
+            dst_ts=ts_dst,
         )
 
     logging.info(